// frontend/src/model/Student.ts
export interface Student {
    registrationNumber: string;
    math?: number;
    literature?: number;
    foreignLanguage?: number;
    physics?: number;
    chemistry?: number;
    biology?: number;
    history?: number;
    geography?: number;
    civicEducation?: number;
    foreignLanguageCode?: string;
}